refactor Hive Coercion code structure to make it more clear
